fix: add Cargo.lock file
According to the [official Rust documentation](https://doc.rust-lang.org/cargo/faq.html#why-have-cargolock-in-version-control), including Cargo.lock in version control is generally not necessary for libraries, as their dependencies are resolved by the consuming application.
However, since we are building a package (not a library), where dependencies are self-contained and cannot be inherited from external sources, maintaining Cargo.lock ensures reproducibility.
This is particularly important when packaging with tools like Nix.